dataprep_trigger/requirements.txt
requests>=2.25.1
google-cloud-storage>=2.0.0